+++ folio/__init__.py
@@ -5,7 +5,7 @@ Python library for FOLIO, the Federated Open Legal Information Ontology
 # SPDX-License-Identifier: MIT
 # (c) 2024 ALEA Institute.
 
-__version__ = "0.1.6"
+__version__ = "0.2.0"
 __author__ = "ALEA Institute"
 __license__ = "MIT"
 __description__ = (
@@ -16,6 +16,13 @@ __url__ = "https://openlegalstandard.org/"
 
 # import graph to re-export
 from .graph import FOLIO, FOLIOTypes, FOLIO_TYPE_IRIS
-from .models import OWLClass, NSMAP
+from .models import OWLClass, OWLObjectProperty, NSMAP
 
-__all__ = ["FOLIO", "FOLIOTypes", "FOLIO_TYPE_IRIS", "OWLClass", "NSMAP"]
+__all__ = [
+    "FOLIO",
+    "FOLIOTypes",
+    "FOLIO_TYPE_IRIS",
+    "OWLClass",
+    "OWLObjectProperty",
+    "NSMAP",
+]